requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pillow
pandas
openpyxl
pydantic
python-dotenv
